correct parsing of , separated vals

+++ block.sh
@@ -31,7 +31,7 @@ TMPFILE=.BLOCK_$$_$RANDOM.tmp
 
 error()
 {
-  rm $TMPFILE
+  if [ -f $TMPFILE ] ; then rm $TMPFILE; fi
   echo "$@" >&2;
   exit 1;
 }
@@ -81,11 +81,10 @@ echo $BLOCK_BEGIN_DELIM > $TMPFILE
 
 #from variable
 if [ "$1" == "-e" ]; then
-  if [ -z $DNS_R2LOCAL_NAMES ]; then
+  if [ -z "$DNS_R2LOCAL_NAMES" ]; then
     error "-e : no environment variable \"DNS_R2LOCAL_NAMES\" found"
   fi
-
-  echo $DNS_R2LOCAL_NAMES | sed "s/,/\n/g" | sed "s/http:\/\///" | sed "s/www.//" | awk '{ print "127.0.0.1 " $1 "\n127.0.0.1 www." $1 }' >> $TMPFILE
+  echo $DNS_R2LOCAL_NAMES | tr , '\n' | sed "s/http:\/\///" | sed "s/www.//" | awk '{ print "127.0.0.1 " $1 "\n127.0.0.1 www." $1 }' >> $TMPFILE
   shift
 fi
 
@@ -97,7 +96,7 @@ if [ "$1" == "-f" ]; then
     error "$2: no such file found"
   fi
 
-  cat $2 | sed "s/,/\n/g" | sed "s/http:\/\///" | sed "s/www.//" | awk '{ print "127.0.0.1 " $1 "\n127.0.0.1 www." $1 }' >> $TMPFILE
+  cat $2 | tr , '\n' | sed "s/http:\/\///" | sed "s/www.//" | awk '{ print "127.0.0.1 " $1 "\n127.0.0.1 www." $1 }' >> $TMPFILE
   shift
   shift
 fi
